// src/modules/reports/interfaces/daily-summary-response.ts
export interface DailyEmployeesSummaryResponse {
  daily_active_employees: number;
  daily_inactive_employees: number;
  daily_suspended_employees: number;
  daily_insured_employees: number;
  daily_accredited_employees: number;
  daily_account_validation_employees: number;
  daily_approval_employees: number;
  daily_pending_employees: number;
  daily_temporary_guard_employees: number;
  daily_training_employees: number;
  daily_under_review_employees: number;
  daily_under_review_iao_employees: number;
  daily_under_review_lic_employees: number;
  daily_under_review_th_employees: number;
}